feat: made the app a functional component
