+++ src/components/loader/Loadet.tsx
@@ -1,14 +1,14 @@
 import { Component } from 'react';
-import './Loader.css'
+import './Loader.css';
 
 class Loader extends Component {
   render() {
     return (
-      <div className="loader">
-        <div className="loader-image">🗘</div>
+      <div className='loader'>
+        <div className='loader-image'>🗘</div>
       </div>
-    )
+    );
   }
 }
 
-export default Loader
+export default Loader;
